+++ oidf-modules/common/src/main/java/se/swedenconnect/oidf/common/entity/entity/integration/InMemoryTrustMarkCache.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2024-2026 Sweden Connect
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package se.swedenconnect.oidf.common.entity.entity.integration;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+
+/**
+ * In-memory implementation of {@link TrustMarkCache}.
+ *
+ * @author Felix Hellman
+ */
+public class InMemoryTrustMarkCache implements TrustMarkCache {
+
+  final Map<Long, Map<String, String>> cache = new ConcurrentHashMap<>();
+
+  @Override
+  public Optional<String> get(final long snapshot, final String trustMarkType, final String subject) {
+    return Optional.ofNullable(this.cache.get(snapshot))
+        .flatMap(version -> Optional.ofNullable(version.get(trustMarkType + ":" + subject)));
+  }
+
+  @Override
+  public void put(final long snapshot, final String trustMarkType, final String subject, final String response) {
+    this.cache.computeIfAbsent(snapshot, _ -> new ConcurrentHashMap<>()).put(trustMarkType + ":" + subject, response);
+  }
+}