+++ oidf-starters/trustmarkissuer-starter/src/main/java/se/swedenconnect/oidf/trustmarkissuer/starter/TrustMarkIssuerRoutingConfiguration.java
@@ -1,23 +1,24 @@
   /*
-   * Copyright 2024-2025 Sweden Connect
-   *
-   * Licensed under the Apache License, Version 2.0 (the "License");
-   * you may not use this file except in compliance with the License.
-   * You may obtain a copy of the License at
-   *
-   *     http://www.apache.org/licenses/LICENSE-2.0
-   *
-   * Unless required by applicable law or agreed to in writing, software
-   * distributed under the License is distributed on an "AS IS" BASIS,
-   * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   * See the License for the specific language governing permissions and
-   * limitations under the License.
-   *
-   */
+ * Copyright 2024-2026 Sweden Connect
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
   package se.swedenconnect.oidf.trustmarkissuer.starter;
 
   import org.springframework.context.annotation.Bean;
   import org.springframework.context.annotation.Configuration;
+  import se.swedenconnect.oidf.common.entity.entity.integration.TrustMarkCache;
   import se.swedenconnect.oidf.common.entity.entity.integration.TrustMarkStatusCache;
   import se.swedenconnect.oidf.common.entity.tree.scraping.ScrapedEntityLookup;
   import se.swedenconnect.oidf.routing.RouteFactory;
@@ -35,7 +36,8 @@
     TrustMarkIssuerRouter trustMarkIssuerRouter(final RouteFactory routeFactory, final TrustMarkIssuerFactory factory,
                                                 final ServerResponseErrorHandler handler,
                                                 final ScrapedEntityLookup lookup,
-                                                final TrustMarkStatusCache trustMarkStatusCache) {
-      return new TrustMarkIssuerRouter(routeFactory, factory, handler, lookup, trustMarkStatusCache);
+                                                final TrustMarkStatusCache trustMarkStatusCache,
+                                                final TrustMarkCache trustMarkCache) {
+      return new TrustMarkIssuerRouter(routeFactory, factory, handler, lookup, trustMarkStatusCache, trustMarkCache);
     }
   }